Creando microservicio-cursos

+++ microservicio-commons/src/main/java/vicente/marti/microserviciocommons/controller/CommonController.java
@@ -9,6 +9,9 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import vicente.marti.microserviciocommons.service.CommonService;
 
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 
 public class CommonController<T, S extends CommonService<T>> {
@@ -32,13 +35,22 @@ public class CommonController<T, S extends CommonService<T>> {
 
     @PostMapping("/")
     public ResponseEntity<?> save(@Validated @RequestBody T type, BindingResult result) {
+        if (result.hasErrors()) return this.validate(result);
         return ResponseEntity.status(HttpStatus.CREATED).body(service.save(type));
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteById(@PathVariable Long id) {
+        Optional<T> optional = service.findById(id);
+        if (optional.isEmpty()) return ResponseEntity.notFound().build();
         service.deleteById(id);
         return ResponseEntity.noContent().build();
+    }
+
+    protected ResponseEntity<?> validate(BindingResult result) {
+        Map<String, Object> errores = new HashMap<>();
+        result.getFieldErrors().forEach(error -> errores.put(error.getField(), "El campo: ".concat(error.getField()).concat(Objects.requireNonNull(error.getDefaultMessage()))));
+        return ResponseEntity.badRequest().body(errores);
     }
 
 }
